Refactoring to remove netlify code and github action for deployment
