+++ .nvm
@@ -1,0 +1,1 @@
+v12.22.12
