--- netlify.toml
@@ -1,2 +1,0 @@
-[build]
-functions = "functions"
