+++ package.json
@@ -2,18 +2,12 @@
   "name": "lily",
   "version": "1.0.0",
   "description": "",
-  "main": "src/server.js",
-  "scripts": {
-    "start": "./node_modules/.bin/netlify-lambda serve src",
-    "build": "./node_modules/.bin/netlify-lambda build src"
-  },
+  "main": "app.js",
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "ejs": "^3.1.7",
-    "express": "^4.18.1",
-    "netlify-lambda": "^2.0.15",
-    "serverless-http": "^3.0.1"
+    "express": "^4.18.1"
   }
 }
